Tailorized "2005-09-30 19:08:15 by logg" Implement compilation for BLAS.
FFC now supports the option -a blas that can be used to generate code for
DOLFIN that uses BLAS calls, instead of generating code that explicitly
computes the tensor products. This has two advantages:

  - The code will run faster for complicated forms
  - The code will compile much faster for complicated forms

Note that it may still take some time to compile the form, but the second
step (gcc) will be lightning fast...

Also note that BLAS is yet only supported on the FFC side. It remains to
implement a few calls on the DOLFIN side (such as reading the data generated
by FFC at startup).

To compile for BLAS, add the option -f blas:

  ffc -f blas NavierStokes.form