Tailorized "2005-10-13 19:54:17 by logg" BLAS mode optimization:
Only assign to entries of G that meet nonzero entries of A in BLAS mode.
Try it out and see if it works. This should only affect the BLAS mode and
it should speed up compile time and run time in cases where there are many entries of
G that meet nonzero entries of A.