Tailorized "2005-11-04 04:57:41 by logg" Work on functionals:
Make sure forms are compiled correctly in cases where the rank
of the element tensor is zero.

This means we are now pretty close to supporting functionals in
FFC/DOLFIN. (Only need to update interfaces on both ends and
implement assembly of scalars.)

Test problem available in src/demo/EnergyNorm.form.