Tailorized "2005-12-05 17:33:55 by logg" Re-add progess bar for computation of reference tensor.
The computation of the reference tensor actually takes about 1 second
for quadratics and 0.1 seconds for linears. Most of the time is now
spent on generating the code (where we have to loop over all entries in
the reference tensor and write them to file).

The computation of the reference tensor is now no longer the bottle-neck.