Refactoring in QuadratureTransformerBase._create_function_name().
In particular focus on making it clearer how the coefficient_access variable
is set depending on different situations. Now it has exactly one initialization
for any given situation. This work made it clearer for me what I need to
do in the corresponding uflacs implementation.

Also renamed quad_element -> is_quad_element in a few files, this got mixed together with the refactoring.