Refactor tabulate_basis slightly and add tables for CG1 elements which are needed for computation of x and J in uflacs.
I'd rather refactor further so that the representation can ask the
table generation for the actual tables it needs rather than generating
a bunch of tables that are probably needed and then vetting them after
the representation is built. But that's a task for later.

Note that this changes the ordering of element tables, which means
a lot of names will change in regression reference code.