Fix issue 100.
Temporary fix for bug where table of jacobian values of high-order
coordinates was compacted because of trailing zeros in precomputed
table. Using formula to compute number of scalar dofs correctly for
Lagrange 1-3 on simplices, which is all dolfin supports for now.
Should rather get the dimension from element tables but those are
so messy and this is sufficient for now.

Refactor jacobian and spatial_coordinate handlers to reuse all the
code, they only differ in which table data they use.

Other minor simplifications.