Make uflacs the default representation.
Rework the auto representation selection a bit,
still need to select 'quadrature' if custom integrals are present,
and want to select 'tensor' if possible for other integrals in the
same form as custom integrals.

Add options --only-auto and --skip-tensor to regression test.py
script, and added tensor as a separate case since it's not covered
by auto anymore, with its own skiplist.