codegeneration/ufc: Add init() and destroy() functions for ufc_finite_element
In this way, the user can pass in a pointer to an object of type
ufc_finite_element, and it can be initialised without relying on the
generated code to call malloc() to obtain storage for the object.

Moreover, the user should call destroy() whenever the object is no
longer needed, so that any implementation-defined resources that are
associated with the object can be freed.